feat: clean server data migrate db logic directly on server actions
