+++ server/actions/priorities.ts
@@ -1,7 +1,7 @@
 'use server'
 
-import { prioritiesMethods as priorities } from '@/server/data/priorities'
+import { database } from '@/server/config/database'
 
 export async function getPriorities() {
-  return await priorities.get()
+  return await database.query.priorities.findMany()
 }
