+++ server/actions/tickets.ts
@@ -1,12 +1,22 @@
 'use server'
 
-import { ticketsMethods as tickets } from '@/server/data/tickets'
+import { eq } from 'drizzle-orm'
+
 import { TicketInformationType } from '@/lib/definitions'
 import { response } from '@/server/lib/response'
+import { database } from '@/server/config/database'
+import schema from '@/server/schema'
 
 export async function getAllTickets() {
   try {
-    const data = await tickets.get()
+    const data = await database.query.tickets.findMany({
+      with: {
+        status: true,
+        priority: true,
+        resolution: true,
+        creator: true
+      }
+    })
     return response(true, 'Tickets fetched successfully', data)
   } catch (error) {
     return response(false, 'An error occurred while fetching tickets')
@@ -15,7 +25,9 @@ export async function getAllTickets() {
 
 export async function getTicketById(id: string) {
   try {
-    const data = await tickets.getById(id)
+    const data = await database.query.tickets.findFirst({
+      where: eq(schema.tickets.id, id)
+    })
     return response(true, 'Ticket fetched successfully', data)
   } catch (error) {
     return response(false, 'An error occurred while fetching a ticket')
@@ -24,7 +36,7 @@ export async function getTicketById(id: string) {
 
 export async function createTicket(form: TicketInformationType) {
   try {
-    const data = await tickets.create(form)
+    const data = await database.insert(schema.tickets).values(form)
     return response(true, 'Ticket created successfully', data)
   } catch (error) {
     return response(false, 'An error occurred while creating a ticket')
