--- server/data/priorities.ts
@@ -1,7 +1,0 @@
-import { database } from '@/server/config/database'
-
-export const prioritiesMethods = {
-  get: async () => {
-    return await database.query.priorities.findMany()
-  }
-}
