--- server/data/roles.ts
@@ -1,7 +1,0 @@
-import { database } from '@/server/config/database'
-
-export const rolesMethods = {
-  get: async () => {
-    return await database.query.roles.findMany()
-  }
-}
