--- server/data/status.ts
@@ -1,7 +1,0 @@
-import { database } from '@/server/config/database'
-
-export const statusMethods = {
-  get: async () => {
-    return await database.query.status.findMany()
-  }
-}
